requirements.txt
antlr4-python3-runtime==4.7.1
boto3==1.9.86
GitPython==2.1.11
pymongo==3.7.2
pytest==4.0.1
pytest-cov==2.6.0
pytest-mock==1.10.0
Flask==1.0.2
Flask-PyMongo==2.2.0
pytest-flask==0.14.0
python-dateutil==2.7.5
requests==2.20.1
pyhcl==0.3.10
